learningflow/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
werkzeug==2.3.7
google-generativeai==0.5.4
python-dotenv==0.21.1
flask-sqlalchemy==3.0.5
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.2
pymysql==1.1.0
cryptography==41.0.3
reportlab==4.0.7